% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_dist.R
\name{get_gamma_parameters}
\alias{get_gamma_parameters}
\alias{gamma_mu}
\alias{gamma_sd}
\title{Gamma Distribution}
\usage{
get_gamma_parameters(population_avg, population_var)

gamma_mu(alpha, beta)

gamma_sd(alpha, beta)
}
\arguments{
\item{population_avg}{the population average, a length 1 numeric value
describing the average (mean) value of the population}

\item{population_var}{the population variance, a length 1 numeric value
describing the variance of the population}

\item{alpha}{alpha parameter of a gamma distribution}

\item{beta}{beta parameter of a gamma distribution}

\item{n}{numeric value representing number of observations}

\item{successes}{numeric value representing the total number of successes across all the n observations.}
}
\description{
Functions for the gamma distribution to calculate priors and
posteriors.
}
\details{
The gamma distribution should be used as a rate per event, the distribution
being predicting is a value between 0 and inf. Think of this as runs/game
like in baseball.
}
\examples{
## Generate data following a beta distribution
library(MASS)

body_log <- log(Animals$body)

gamma_mean <- mean(body_log)
gamma_var <- var(body_log)

## Using the function
gamma_parameters <- get_gamma_parameters(
  population_avg = gamma_mean,
  population_var = gamma_var
)

gamma_parameters

}
