% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-writing.R
\name{guess_write_function}
\alias{guess_write_function}
\title{Guess write function from file extension}
\usage{
guess_write_function(file_name)
}
\arguments{
\item{file}{filename to parse}
}
\value{
function for reading the file, if found
}
\description{
This is a slight modification to the guess_write function from the piggyback paackage
}
\details{
\code{guess_write_function} understands the following file extensions:
\itemize{
\item csv, csv.gz, csv.xz with \code{utils::read.csv}
\item parquet with \code{arrow::read_parquet}
}
}
\keyword{internal}
