% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{import_palmer_penguins}
\alias{import_palmer_penguins}
\title{Importing from s3 bucket}
\usage{
import_palmer_penguins(write_to_directory = FALSE, path = NULL, country = NULL)
}
\arguments{
\item{write_to_directory}{controls whether data is written to local directory. Defaults to \code{TRUE}}

\item{path}{name of directory you would like to write the dataset to}

\item{country}{If you want a subset of the data please provide a string for something coercible to a string.}
}
\description{
This function imports our data from an s3 AWS bucket
}
