% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteria.R
\name{InfCriteria}
\alias{InfCriteria}
\title{Calculate Information Criteria Value for Model Selection}
\usage{
InfCriteria(loglikelihood, clusters, dimensions, nObservations = 100)
}
\description{
Calculates information criteria values, BIC and AIC for model
selection in Gaussian mixture models, provided loglikelihood
number of clusters, dimensionality of dataset, and number
of observations
}
\examples{
InfCriteriaV (loglikelihood = -5080,
               clusters = 2,
               dimensions = 3,
              nObservations = 1000)


}
