% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteria_plot.R
\name{InfCriteria_plot}
\alias{InfCriteria_plot}
\title{Plot Information Criteria Values}
\usage{
InfCriteria_plot(inf_values)
}
\arguments{
\item{inf_values}{A list of negative values of class "InfCriteria" indicating
the log-likelihood corresponding to BIC, AIC and ICL, respectively.}
}
\value{
Returns a plot of information criteria values
}
\description{
A function that plots information criteria values
}
\examples{
InfCriteria_results <- InfCriteria_v3(loglikelihood = -5080,
                                      clusters = 2,
                                      dimension = 3,
                                      observations = 1000,
                                      probability = c(0.5, 0.5))
InfCriteria_plot(inf_values = InfCriteria_results)

}
