% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteria_calculation.R
\name{InfCriteria_v3}
\alias{InfCriteria_v3}
\title{Calculates Information Criteria}
\usage{
InfCriteria_v3(loglikelihood, clusters, dimension, observations, probability)
}
\arguments{
\item{loglikelihood}{A negative value of class "numeric" indicating
the log-likelihood}

\item{clusters}{A positive integer indicating the number of clusters}

\item{dimension}{A positive integer indicating the dimension of dataset}

\item{observations}{A positive integer indicating the number of observations}

\item{probability}{A vector indicating the probability of each cluster}
}
\value{
Returns an S3 object of class InfCriteria with results.
\itemize{
\item BICresults - A value of class "numeric" indicating BIC value
\item AICresults - A value of class "numeric" indicating AIC value
\item ICLresults - A value of class "numeric" indicating ICL value
}
}
\description{
A function that calculates information criteria given log-likelihood,
number of clusters, dimension of dataset, and number of observations,
and the probability
}
\examples{
InfCriteria_results <- InfCriteria_v3(loglikelihood = -5080,
                                      clusters = 2,
                                      dimension = 3,
                                      observations = 1000,
                                      probability = c(0.5, 0.5))
InfCriteria_results$BICresults

}
\references{
Akaike, H. (1973). Information theory and an extension of the maximum
likelihood principle. In \emph{Second International Symposium on Information
Theory}, New York, NY, USA, pp. 267–281. Springer Verlag. \href{https://link.springer.com/chapter/10.1007/978-1-4612-1694-0_15}{Link}
}
