% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.R
\name{multiple}
\alias{multiple}
\title{Multiples of a given number}
\usage{
multiple(x)
}
\arguments{
\item{x}{A number to get the multiple for}
}
\value{
a vector of multiple of given number
}
\description{
The function generates the first 10 multiples of a given number
}
