% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTestingPackage.R
\name{runTestingPackage}
\alias{runTestingPackage}
\title{Launch the shiny app for package TestingPackage}
\usage{
runTestingPackage()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for TestingPackage.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runTestingPackage()
}

}
