% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate input parameters}
\usage{
validate_inputs(write_to_directory, path, country, file_name, write_function)
}
\description{
keywords @internal
}
