% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteriaCalculation.R
\name{InfCriteriaCalculation}
\alias{InfCriteriaCalculation}
\title{Calculates Information Criteria}
\usage{
InfCriteriaCalculation(
  loglikelihood,
  nClusters = 2,
  dimensionality,
  observations,
  probability
)
}
\arguments{
\item{loglikelihood}{A negative value of class "numeric" indicating
the log-likelihood.}

\item{nClusters}{A positive integer indicating the number of clusters.
Default value is 2.}

\item{dimensionality}{A positive integer indicating the dimensionality of dataset.}

\item{observations}{A positive integer indicating the number of observations.}

\item{probability}{A vector indicating the probability of each cluster.}
}
\value{
Returns an S3 object of class InfCriteria with results.
\itemize{
\item BICresults - A value of class "numeric" indicating BIC value.
\item AICresults - A value of class "numeric" indicating AIC value.
\item ICLresults - A value of class "numeric" indicating ICL value.
}
}
\description{
A function that calculates information criteria given log-likelihood,
number of clusters, dimension of dataset, and number of observations,
and the probability.
}
\examples{
# Examples 1:
# Using GeneCounts dataset available with package
dim(GeneCounts)

# Calculate information criteria value
InfCriteriaResults <- InfCriteriaCalculation(
                          loglikelihood = -5080,
                          nClusters = 2,
                          dimensionality = ncol(GeneCounts),
                          observations = nrow(GeneCounts),
                          probability = c(0.5, 0.5))
InfCriteriaResults$BICresults

# Examples 2:
# Obtain an external sample RNAseq dataset
\dontrun{
library(MBCluster.Seq)
data("Count")
dim(Count)

# Calculate information criteria value
InfCriteriaResults <- InfCriteriaCalculation(
                          loglikelihood = -5080,
                          nClusters = 2,
                          dimensionality = ncol(Count),
                          observations = nrow(Count),
                          probability = c(0.5, 0.5))
InfCriteriaResults$BICresults
}

# Example 3:
# Another way to obtain external data from another
# package that is only used for the example:
# More info, see: https://r-pkgs.org/r-cmd-check.html

if (requireNamespace("RTCGA.rnaseq", quietly = TRUE)) {
library(RTCGA.rnaseq)
dim(ACC.rnaseq) # 79 20532

# Calculate information criteria value
InfCriteriaResults <- InfCriteriaCalculation(
                          loglikelihood = -5080,
                          nClusters = 2,
                          dimensionality = nrow(ACC.rnaseq),
                          observations = ncol(ACC.rnaseq),
                          probability = c(0.5, 0.5))
InfCriteriaResults$BICresults


# Examples 4:
# Access raw data made available with this package
inputCountsPath <- system.file("extdata", "GeneCounts.csv", package = "TestingPackage")
# Read data
inputCounts <- read.csv(file = inputCountsPath, header = TRUE)
dim(inputCounts) # 30  4

# To see documentation for data, made available seperately
?GeneCounts

# Calculate information criteria value
InfCriteriaResults <- InfCriteriaCalculation(
                          loglikelihood = -5080,
                          nClusters = 2,
                          dimensionality = nrow(inputCounts),
                          observations = ncol(inputCounts),
                          probability = c(0.5, 0.5))
InfCriteriaResults$BICresults
}

}
\references{
Akaike, H. (1973). Information theory and an extension of the maximum
likelihood principle. In \emph{Second International Symposium on Information
Theory}, New York, NY, USA, pp. 267–281. Springer Verlag. \href{https://link.springer.com/chapter/10.1007/978-1-4612-1694-0_15}{Link}

Biernacki, C., G. Celeux, and G. Govaert (2000). Assessing a mixture model for
clustering with the integrated classification likelihood. \emph{IEEE Transactions on Pattern
Analysis and Machine Intelligence} 22. \href{https://hal.inria.fr/inria-00073163/document}{Link}

Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics} 6, 461–464.
\href{https://projecteuclid.org/euclid.aos/1176344136}{Link}.
}
