% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteriaCalculation.R
\name{InfCriteriaCalculation}
\alias{InfCriteriaCalculation}
\title{Calculates Information Criteria}
\usage{
InfCriteriaCalculation(
  loglikelihood,
  clusters,
  dimension,
  observations,
  probability
)
}
\arguments{
\item{loglikelihood}{A negative value of class "numeric" indicating
the log-likelihood}

\item{clusters}{A positive integer indicating the number of clusters}

\item{dimension}{A positive integer indicating the dimension of dataset}

\item{observations}{A positive integer indicating the number of observations}

\item{probability}{A vector indicating the probability of each cluster}
}
\value{
Returns an S3 object of class InfCriteria with results.
\itemize{
\item BICresults - A value of class "numeric" indicating BIC value
\item AICresults - A value of class "numeric" indicating AIC value
\item ICLresults - A value of class "numeric" indicating ICL value
}
}
\description{
A function that calculates information criteria given log-likelihood,
number of clusters, dimension of dataset, and number of observations,
and the probability
}
\examples{
InfCriteriaResults <- InfCriteriaCalculation(loglikelihood = -5080,
                                             clusters = 2,
                                             dimension = 3,
                                             observations = 1000,
                                             probability = c(0.5, 0.5))
InfCriteriaResults$BICresults

}
\references{
Akaike, H. (1973). Information theory and an extension of the maximum
likelihood principle. In \emph{Second International Symposium on Information
Theory}, New York, NY, USA, pp. 267–281. Springer Verlag. \href{https://link.springer.com/chapter/10.1007/978-1-4612-1694-0_15}{Link}

Biernacki, C., G. Celeux, and G. Govaert (2000). Assessing a mixture model for
clustering with the integrated classification likelihood. \emph{IEEE Transactions on Pattern
Analysis and Machine Intelligence} 22. \href{https://hal.inria.fr/inria-00073163/document}{Link}

Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics} 6, 461–464.
\href{https://projecteuclid.org/euclid.aos/1176344136}{Link}.
}
