% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteriaPlot.R
\name{InfCriteriaPlot}
\alias{InfCriteriaPlot}
\title{Scatter Plot Of Information Criteria Values}
\usage{
InfCriteriaPlot(infValues, plotTitle = NA)
}
\arguments{
\item{infValues}{A list of positive (double or integer) values
of class "InfCriteria" indicating information criteria values
corresponding to BIC, AIC and ICL, respectively.}

\item{plotTitle}{A character vector specifying a custom name
for the plot to be created. Default is "Plot of information
criteria values".}
}
\value{
Returns a scatter plot showing information criteria values
versus the criterion.
}
\description{
A function that plots information criteria values as a scatter plot.
Here, information criterion AIC, BIC and ICL are provided along x-axis
and information criterion value along the y-axis.
}
\examples{
InfCriteriaResults <- TestingPackage::InfCriteriaCalculation(
                                             loglikelihood = -5080,
                                             nClusters = 2,
                                             dimensionality = 3,
                                             observations = 1000,
                                             probability = c(0.5, 0.5))
InfCriteriaPlot(infValues = InfCriteriaResults)

}
