% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteriaPlot.R
\name{InfCriteriaPlot}
\alias{InfCriteriaPlot}
\title{Plot Information Criteria Values}
\usage{
InfCriteriaPlot(infValues)
}
\arguments{
\item{infValues}{A list of negative values of class "InfCriteria" indicating
the log-likelihood corresponding to BIC, AIC and ICL, respectively.}
}
\value{
Returns a plot of information criteria values
}
\description{
A function that plots information criteria values
}
\examples{
InfCriteriaResults <- InfCriteriaCalculation(loglikelihood = -5080,
                                             nClusters = 2,
                                             dimensionality = 3,
                                             observations = 1000,
                                             probability = c(0.5, 0.5))
InfCriteriaPlot(infValues = InfCriteriaResults)

}
