% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFactorCalculation.R
\name{NormFactors}
\alias{NormFactors}
\title{Calculate Norm Factors}
\usage{
NormFactors(dataset)
}
\arguments{
\item{dataset}{A matrix with n observations along rows and d dimensions.}
}
\value{
Returns an object, norm_factors, with normalization factors.
}
\description{
A function that calculate normalization factors via TMM given a dataset
}
\examples{
  dim(GeneCounts)
  normGeneCounts <- NormFactors(dataset = GeneCounts)

}
