% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfCriteriaCalculation.R
\name{NormFactors}
\alias{NormFactors}
\title{Calculate Normalization Factors}
\usage{
NormFactors(dataset)
}
\arguments{
\item{dataset}{A matrix of integers, with n observations along rows
and d dimensions for columns. Typically, d < n.}
}
\value{
Returns an object, normFactors, with normalization factors.
}
\description{
A function that calculate normalization factors via trimmed mean of
M-values (TMM) given an RNA sequencing dataset containing raw counts.
}
\examples{
dim(GeneCounts)
normGeneCounts <- NormFactors(dataset = as.matrix(GeneCounts))

}
\references{
Robinson, M. D., McCarthy, D. J. and Smyth, G. K. (2010). edgeR: a
Bioconductor package for differential expression analysis of digital gene
expression data. \emph{Bioinformatics}, 26, 139-140.
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2796818/}{Link}

McCarthy, D. J., Chen, Y. and Smyth, G. K. (2012). Differential expression
analysis of multifactor RNA-Seq experiments with respect to biological
variation. \emph{Nucleic Acids Research}, 40, 4288-4297.
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3378882/}{Link}
}
