% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormFactor_calculation.R
\name{NormFactors}
\alias{NormFactors}
\title{Calculate Norm Factors}
\usage{
NormFactors(dataset)
}
\arguments{
\item{dataset}{A negative value of class "numeric" indicating
the log-likelihood}
}
\value{
Returns an object, norm_factors, with normalization factors.
}
\description{
A function that calculate normalization factors via TMM
}
\examples{
  dim(GeneCounts)
  norm_factors <- log(as.vector(edgeR::calcNormFactors(as.matrix(GeneCounts), method = "TMM")))

}
